+++ CHANGELOG.md
@@ -5,7 +5,6 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
-- 'Journal' theme.
 - 'Litera' theme.
 - 'Lumen' theme.
 - 'Sandstone' theme.
@@ -14,12 +13,23 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/) and this 
 - 'Spacelab' theme.
 - 'Superhero' theme.
 
+## [0.1.2] - 2020-05-10
+### Added
+- Browser detection to enable dark mode
+- 'Journal' theme.
+
+### Changed
+- Minor changes to prior themes
+- Help page
+
+
 ## [0.1.1] - 2020-05-06
 ### Added
 - API
 
 ### Changed
 - Help page
+
 
 ## [[0.1.0]](https://cdn.jsdelivr.net/gh/vinorodrigues/foundswatch@0.1.0/) - 2020-05-03
 ### Added
@@ -35,6 +45,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/) and this 
 	- materia
 	- united
 
-# Help
+
+## Help
 
 Visit [foundswatch.com/help](https://foundswatch.com/help/).